requirements.txt
click @ file:///build/click-8.0.3/dist/click-8.0.3-py3-none-any.whl
pycryptodome @ file:///build/pycryptodome-3.11.0/dist/pycryptodome-3.11.0-cp35-abi3-linux_x86_64.whl
python-dotenv @ file:///build/python-dotenv-0.19.0/dist/python_dotenv-0.19.0-py2.py3-none-any.whl
python-vlc @ file:///build/python-vlc-3.0.12118/dist/python_vlc-3.0.12118-py3-none-any.whl
youtube-dl @ file:///build/youtube-dl/dist/youtube_dl-2021.12.17-py2.py3-none-any.whl
